requirements.txt
numpy
pandas
matplotlib
pyexiftool
